requirements.txt
python-telegram-bot==21.9
APScheduler==3.10.4
pytz
tzlocal
gspread
google-auth